Extract AggregateScanState to its own file.
